% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSalary.R
\name{getSalary}
\alias{getSalary}
\title{Retrieve salary values (if present) from the Indeed API.}
\usage{
getSalary(url, currency = c("USD", "GBP", "EUR"))
}
\arguments{
\item{url}{(string). The url of the job description page. Required. No
default.}

\item{currency}{(string). Currency of the salary. The default is USD ($).
Currently, the only supported currencies are USD, GBP and EUR.}
}
\value{
A dataframe is returned, consisting of:
  status: Job status (contract, permanent, temporary);
  period: Salary period (day, week, month, year);
  currency: The currency passed as an argument;
  minSal: If a range is detected (e.g. $40,000 - $60,000), then this is the
  lower value. If no range is dectected, then minSal and maxSal will equal this value;
  maxSal: If a range is detected (e.g. $40,000 - $60,000), then this is the
  higher value. If no range is dectected, then minSal and maxSal will equal this value.
}
\description{
Salary figures are not explicity returned by the Indeed API (see
\code{jobSearch} and \code{jobkeySearch}). \code{getSalary} tries to
overcome this by scraping any salary figures from the job description page
(the results.url column from the \code{jobSearch} or \code{jobkeySearch}
output). Note that it's quite common for job descriptions to not include any
salary data, which will return NA.
}
\examples{
\dontrun{
# search for data scientist jobs in Dublin, Ireland (10 jobs returned)
jobs <- searchJobs(publisher="123445678", query="data+scientist",
                   country = "ie", location = "dublin")

# retrieve salary for the first job posted
getSalary(jobs$results.url[1],"GBP")

# retrieve salary for all jobs
lapply(jobs$results.url,function(x){getSalary(x,"GBP")})

}

}

