% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indeedCountryCodes}
\alias{indeedCountryCodes}
\title{Indeed Country Codes}
\format{A data frame with 2 variables: \code{country} and
  \code{code}}
\usage{
indeedCountryCodes
}
\description{
The country codes for the Indeed API (e.g. "us" is the code for the United States of America)
}
\keyword{datasets}

