% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobkeySearch.R
\name{jobkeySearch}
\alias{jobkeySearch}
\title{Perform job search by job key}
\usage{
jobkeySearch(publisher, jobkeys, version = 2)
}
\arguments{
\item{publisher}{(string) Publisher ID (e.g. "901314124155123"), which is
assigned when you register as a publisher. No default. (required)}

\item{jobkeys}{(sting <vector>). Job keys. Vector of job keys specifying the
jobs to look up. This parameter is required. No default.}

\item{version}{(integer). Specifies the version of the API you wish to use.
All publishers should be using version 2. Currently available versions are
1 and 2. Default is 2.}
}
\value{
The output is converted from its JSON format to a data frame.
}
\description{
Performs a job search by sending job keys to the Indeed API. The output
fields are identical to \code{jobSearch}, except that the response does not
include the various search parameters, such as query or radius.
}
\examples{
\dontrun{
# returns information for the two provided jobkeys
jobkeySearch(publisher="123456789", c("6e9c05292f3275f5","031c1652b9692d2e"))
}

}

