% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobSearch.R
\name{jobSearch}
\alias{jobSearch}
\title{Perform simple job search}
\usage{
jobSearch(publisher, query, country = "us", location = "", radius = 25,
  sort = c("relevance", "date"), start = 0, limit = 10, all = FALSE,
  latlong = "", fromAge = "", st = "", jt = "", filter = "",
  userip = "1.2.3.4", version = 2,
  userAgent = "Mozilla/\%2F4.0\%28Firefox\%29", callback = "",
  highlight = "", chnl = "")
}
\arguments{
\item{publisher}{(string) Publisher ID (e.g. "901314124155123"), which is assigned when you register as a publisher. No default. (required)}

\item{query}{(string) Job query. For simple queries, words are seperated by a plus sign e.g. "data+scientist". For more advanced queries, use the adavanced search \url{http://www.indeed.co.uk/advanced_search} and check the corresponding value of q in the url. No default. (required)}

\item{country}{(string). Specifies the country of the job search. The default is "us", which corresponds to the USA.}

\item{location}{(string). Specific area within country over which you want to perform the job search. Accepts postal codes or a  "city, state/province/region" combination (see \url{https://ads.indeed.com/jobroll/xmlfeed} for country specific codes). Default is "", which returns jobs for the whole country.}

\item{radius}{(integer). Allowable distance in miles from search location ("as the crow flies"). Default is 25.}

\item{sort}{(string). Sort by relevance or date. Default is "relevance".}

\item{start}{(integer). Start results at this result number. Default is 0.}

\item{limit}{(integer). Maximum number of results returned per query (1-25). Default is 10.}

\item{all}{(boolean). Single API calls are limited to 25 jobs. To return all jobs in a single data frame, then set all = TRUE (this will override the start and limit arguments). Default is all = FALSE.}

\item{latlong}{(integer). If latlong=1, search returns latitude and longitude information for each job result. Default is 0.}

\item{fromAge}{(integer). Number of days back to search. Default is currently unknown to me.}

\item{st}{(string). Site type. To show only jobs from job boards use "jobsite". For jobs from direct employer websites, use "employer". Default is "" i.e. show all jobs.}

\item{jt}{(string). Job type. Allowed values: "fulltime", "parttime", "contract", "internship", "temporary". Default is "" i.e. no job type distinction within search.}

\item{filter}{(integer). Filter duplicate results. 0 turns off duplicate job filtering. Default is 1.}

\item{userip}{(string). The IP number of the end-user to whom the job results will be displayed. The default value is "1.2.3.4", which should cover most cases. Unless you use the wrapper for some front end work, there's no need to alter the default value.}

\item{version}{(integer). Specifies the version of the API you wish to use. All publishers should be using version 2. Currently available versions are 1 and 2. Default is 2.}

\item{userAgent}{(string). The User-Agent (browser) of the end-user to whom the job results will be displayed. The default value is "Mozilla/\%2F4.0\%28Firefox\%29", which should cover most cases. Unless you use the wrapper for some front end work, there's no need to alter the default value.}

\item{callback}{(string).  The name of a javascript function to use as a callback to which the results of the search are passed. For security reasons, the callback name is restricted letters, numbers, and the underscore character. The default value is "", which should cover most cases. Unless you use the wrapper for some front end work, there's no need to alter the default value.}

\item{highlight}{(integer). Setting this value to 1 will bold terms in the snippet that are also present in q. Default is 0.}

\item{chnl}{(string). Channel Name. Group API requests to a specific channel.}
}
\value{
The output is converted from its JSON format to a data frame.
}
\description{
Performs a simple job search call to the Indeed API.
}
\examples{
\dontrun{
# search for data scientist jobs in Dublin, Ireland (10 jobs returned)
searchJobs(publisher="123445678", query="data+scientist",
           country = "ie", location = "dublin")

# search for all data scientist jobs in Dublin, Ireland, sorted by date
searchJobs(publisher="123445678", query="data+scientist",
           country = "ie", location = "dublin", all = TRUE, sort="date")

}

}

